fix: some refactor about dataZoom and axis scale extent calculation:
(1) The "scale extent calculation" is depended by both coord sys and "dataZoom".
So it need to be performed in different stage of workflow and should avoid to call it repleatly or implement it repeatly.
But previously, the "scale extent calculation" is implemented both in `axisHelper.ts` and `dataZoom/AxisProxy.ts`,
where different code implements similar logic.
"Scale extent calculation" is based on input of:
    (I) option specified min/max/scale (including callback) and
    (II) dataExtent and
    (III) some filter related components like dataZoom.
Currently we need add more filters depends on that axis scale calculation.
e.g.,
    (I) one axis min max effect the other axis extent
    (II) custom filter.
So we refactor that "Scale extent calculation" as a reusable module (see `scaleRawExtentInfo.ts`).

(2) Based on (1), make the callback of axis.min/axis.max consistent whatever it is called in
"data processing stage" or "coord sys updating stage"

(3) Based on (1), fix some "inconsistent" flaw in dataZoom when handling stacked series.

(4) Fix the type of axis.min/max to `ScaleDataValue`. And enable the callback of axis.min/max to return `ScaleDataValue`.

(5) Remove some code that never used.

(6) Make the "data filter", which will "block stream", not as the default data processor in the register API.